% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_connection.R
\name{createConnection}
\alias{createConnection}
\title{createConnection: Create a database connection object}
\usage{
createConnection(
  hostname = "localhost",
  port = 1523,
  sid = "CasRef01",
  drv_path = "inst/java/ojdbc7.jar",
  username = NA
)
}
\arguments{
\item{hostname}{Name of host machine.  Default is "localhost".  When connecting from CAS3: "cas2.asckey"}

\item{port}{Port number of host machine.  Default is 1523.  When connecting from CAS3: 1521}

\item{sid}{Database (schema) ID.  Default is "CasRef01"}

\item{drv_path}{Path and file name of JDBC driver.  Default is "ojdbc7.jar"}

\item{username}{CAS username}
}
\value{
Returns a database connection object.
}
\description{
createConnection: Create a database connection object
}
\examples{
\dontrun{
casref01 = createConnection()
}
}
