% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_connection.R
\name{dbExplainPlan}
\alias{dbExplainPlan}
\title{Requests an explain plan for a query in an Oracle database and prints it.}
\usage{
dbExplainPlan(connection, querystring, ..., returnplan = FALSE)
}
\arguments{
\item{connection}{JDBCConnection object created separately.}

\item{querystring}{Query or statement to execute, not terminated by a semicolon unless a PL/SQL statement.}

\item{...}{Arguments to fill in holes in querystring if querystring is a prepared statement.}

\item{returnplan}{Default FALSE, set to TRUE to return the plan table as a data.frame of text.}
}
\value{
None
}
\description{
Requests an explain plan for a query in an Oracle database and prints it.
}
\examples{
\dontrun{
dbGetQueryOracle(con, query)
dbSendQueryOracle(con, statement)
dbGetQueryOracle(con, statement, returndata = FALSE)
dbGetQueryOracle(con, "select ? from dual union select ? from dual", "cow", "hat", timeit = TRUE)
}

}
