% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle_connection.R
\name{dbExploreOracle}
\alias{dbExploreOracle}
\title{Builds a wrapper for dbGetQuery for exploration queries}
\usage{
dbExploreOracle(connection, rowlimit = 50, timeit = FALSE)
}
\arguments{
\item{connection}{JDBCConnection object created separately.}

\item{rowlimit}{Default 50, number of rows queries should return}

\item{timeit}{Default FALSE, turn to TRUE to enable timing.}
}
\value{
None
}
\description{
Builds a wrapper for dbGetQuery for exploration queries
}
\examples{

\dontrun{
explorer = dbExploreOracle(con, rowlimit = 10)
explorer("select * from dual")
}

}
