% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{ndrs_prevalence}
\alias{ndrs_prevalence}
\title{NDRS Prevalence Function}
\usage{
ndrs_prevalence(
  numerator_data,
  denominator_data,
  end_date,
  start_date = NULL,
  years_before_point = NULL,
  conganom_data = FALSE,
  rare_diseases_data = FALSE,
  cancer_data = FALSE,
  per_people = NULL,
  conganom_prev_type = NULL,
  write_csv = FALSE
)
}
\arguments{
\item{numerator_data}{A dataframe with row level patient data.}

\item{denominator_data}{A dataframe with population count per year, lsoa11 and age.}

\item{end_date}{Character string with date in format YYYY-MM-DD. Marks the point prevalence date.}

\item{start_date}{Character string with date in format YYYY-MM-DD. Use for subsetting the data over a period of time.}

\item{years_before_point}{Integer to set number of years before end date to subset the data over a period of time.}

\item{conganom_data}{Logical. Set to TRUE if using congenital anomalies data, default is FALSE.}

\item{rare_diseases_data}{Logical. Set to TRUE if using rare diseases data, default is FALSE.}

\item{cancer_data}{Logical. Set to TRUE if using cancer data, default is FALSE.}

\item{per_people}{Integer for prevalence per x people. Defaults to 10,000 for congenital anomalies, 100,000 for cancer and 1,000,000 for rare diseases.}

\item{conganom_prev_type}{Character string to select prevalence calculation type for congenital anomalies. Accepts 'point', 'livebirth' or 'totalbirth' and default is set to 'point'.}

\item{write_csv}{Logical. Set to TRUE to download a csv copy of output to working directory, default is FALSE.}
}
\value{
A data frame with a row per condition / disease, and columns providing the name, code, prevalence rate, confidence intervals and time period.
A .csv copy of the dataframe is also outputted to the working directory if write_csv is set to TRUE.
}
\description{
Calculating prevalence using raw data from the National Disease Registration Service.
}
\details{
This functions requires 2 data frames: patient level numerator data and raw ONS populations data for the denominator.
An end date is also required to calculate the point prevalence and there are further options such as to calculate livebirth or total birth prevalence for congenital anomalies.
}
\examples{
# Using raw data from rare diseases to calculate the point prevalence
ndrs_prevalence(numerator_data = rare_diseases_raw,
                denominator_data = ons_mid_year_population_data,
                end_date = '2020-12-31',
                rare_diseases_data = TRUE)

# Using raw cancer data to calculate 20 year prevalence
ndrs_prevalence(numerator_data = rare_cancer_data,
                denominator_data = ons_mid_year_population_data,
                end_date = '2020-06-01',
                years_before_point = 20,
                cancer_data = TRUE)

# Using raw data from congenital anomalies to calculate the livebirth prevalence
ndrs_prevalence(numerator_data = congenital_anomalies_raw,
                denominator_data = birth_population_data,
                end_date = '2020-12-31',
                conganom_data = TRUE,
                per_people = 10000,
                conganom_prev_type = 'livebirth',
                write_csv = TRUE)
}
